updated bridge module files structure
