+++ src/modules/bridge/bridge-form/InfoPanel.jsx
